backend/requirements.txt
# FastAPI Framework
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# OCR & Image Processing
Pillow==11.3.0
pdf2image==1.17.0
opencv-python-headless==4.12.0.88
numpy>=2.0.0,<2.3.0
azure-cognitiveservices-vision-computervision==0.9.0
requests==2.32.5

# Excel/CSV Processing
openpyxl==3.1.5
pandas==2.3.3
xlrd==2.0.2

# PDF Processing
PyPDF2==3.0.1
reportlab==4.4.4

# Email
python-dotenv==1.1.1
aiosmtplib==4.0.2
jinja2==3.1.6

# Validation
pydantic==2.11.10
pydantic-settings==2.11.0
email-validator==2.3.0

# Date/Time
python-dateutil==2.9.0.post0
pytz==2025.2

# HTTP Requests
httpx==0.28.1
aiohttp==3.13.0

# Utilities
python-slugify==8.0.4
qrcode[pil]==8.2

# Testing
pytest==8.4.2
pytest-asyncio==1.2.0

# CORS
fastapi-cors==0.0.6

# Redis (opcional)
redis==6.4.0

# Logging
loguru==0.7.3
